# perf: 🤖  remove unused field, better parallel resolve