--- a/crates/swc_ecma_minifier/src/pass/unique_scope.rs
+++ b/crates/swc_ecma_minifier/src/pass/unique_scope.rs
@@ -13,7 +13,7 @@ struct UniqueScope;
 
 impl UniqueScope {
     fn make_unique(&self, span: &mut Span) {
-        span.ctxt = span.ctxt.apply_mark(Mark::fresh(Mark::root()));
+        span.ctxt = span.ctxt.apply_mark(Mark::fresh());
     }
 }