… .vscode/settings.json
… crates/swc_ecma_minifier/src/cli/bin.rs
… crates/swc_ecma_minifier/src/compress/optimize/iife.rs
… crates/swc_ecma_minifier/src/metadata/tests.rs
… crates/swc_ecma_minifier/tests/mangle.rs
… crates/swc_ecma_preset_env/benches/polyfills.rs
… crates/swc_ecma_preset_env/data/core-js-compat/data.json
… crates/swc_ecma_preset_env/data/core-js-compat/entries.json
… crates/swc_ecma_preset_env/data/core-js-compat/modules-by-versions.json
… crates/swc_ecma_preset_env/tests/test.rs
… crates/swc_ecma_transforms/tests/decorators.rs
… crates/swc_ecma_transforms_base/src/helpers/mod.rs
… crates/swc_ecma_transforms_base/src/hygiene/tests.rs
… crates/swc_ecma_transforms_base/src/resolver/tests.rs
… crates/swc_ecma_transforms_base/tests/fixture.rs
… crates/swc_ecma_transforms_classes/src/super_field.rs
… crates/swc_ecma_transforms_compat/src/es2015/block_scoping/mod.rs
… crates/swc_ecma_transforms_compat/src/es2015/block_scoping/vars.rs
… crates/swc_ecma_transforms_compat/src/es2015/classes/mod.rs
… crates/swc_ecma_transforms_compat/src/es2015/computed_props.rs
… crates/swc_ecma_transforms_compat/src/es2015/for_of.rs
… crates/swc_ecma_transforms_compat/src/es2015/mod.rs
… crates/swc_ecma_transforms_compat/src/es2015/parameters.rs
… crates/swc_ecma_transforms_compat/src/es2018/object_rest_spread.rs
… crates/swc_ecma_transforms_compat/src/es2022/class_properties/mod.rs
… crates/swc_ecma_transforms_compat/src/es2022/private_in_object.rs
… crates/swc_ecma_transforms_compat/tests/es2015_block_scoping.rs
… crates/swc_ecma_transforms_compat/tests/es2015_classes.rs
… crates/swc_ecma_transforms_compat/tests/es2015_generator.rs
… crates/swc_ecma_transforms_compat/tests/es2017_async_to_generator.rs