@IWANABETHATGUY 2022-09-30
> If you are trying to fix the perf issue of `resolver`, I expect this to make the situation worse

Would you mind giving me some advice?