@kdy1 2022-09-30
You can optimize it by allocating more marks at once and storing it in a thread-local