@IWANABETHATGUY 2022-10-01
> You can optimize it by allocating more marks at once and storing it in a thread-local

Thanks, I found another way to optimize `Mark::new()`, give me some time to explain about it.