@IWANABETHATGUY 2022-10-01
Here is the `swc_ecma_transforms_base/bench/parallel/ result: 
# main
```
es/base/parallel/resolver/typescript                                                                          
                        time:   [13.508 s 13.631 s 13.743 s]
Found 1 outliers among 10 measurements (10.00%)
```

# PR
```
es/base/parallel/resolver/typescript                                                                          
    time:   [12.448 s 12.577 s 12.722 s]
    change: [-11.218% -9.6772% -7.9568%] (p = 0.00 < 0.05)
    Performance has improved.
```


About 10%, not bad.