@IWANABETHATGUY 2022-10-02
@kdy1 thanks for your reviewing, I close this pull request and make another pr https://github.com/swc-project/swc/pull/6019/files, with no API deletion, only adjusting the data structures.